Store synapse's database somewhere other than a volume
/data is a volume in the Synapse docker image. Volumes are not copied
over when committing a container to an image. Thus, we should store the
database somewhere else.

An alternative fix is to just not create a volume (we don't need it
when running complement tests), however this would require creating an
separate dockerfile from the one we're using from the Synapse repo now,
which is much more annoying.