Address some review feedback
I tried to split this into multiple commits but VSCode tried to delete my
commits instead.

*Anyhoo*, it took me a while to realise, but it turns out that leave events
don't come down /sync unless you're doing an incremental sync. So the
majority of changes are for performing an initial sync, and then saving
the next_batch token to use with SyncUntil.

Otherwise there was a typo fix, inlining the knockingUserID var, and
using must.NotError instead of manual err checking and logging.