Add tests for knocking (#44)
* Add a blueprint for a single, two-user homeserver

* knocking wip

* Add method SyncUntilKnockHas

This method blocks until an entry for a given room appears in the
knock section of /sync

* Use SyncUntil directly for knocking tests

This change removes SyncUntilKnockHas and uses SyncUntil directly instead. This
requires making SyncUntil public, as well as GjsonEscape.

* use MSC number where applicable

* remove room state checking, add ban->knock deny check

the required state events is implementation-specific, so it will need to be
moved to unit tests.

* Test knocking over federation

This needs some cleanup, but on the whole it works!

* minor cleanup

* Add test for user knocking without a reason

* Ensure we ban the user in our ban test

* Add test for knocking once you're already in the room

* more cleanup

* Don't forget to attach our blueprints

* Use SendEventSynced when sending state

* Address some review feedback

I tried to split this into multiple commits but VSCode tried to delete my
commits instead.

*Anyhoo*, it took me a while to realise, but it turns out that leave events
don't come down /sync unless you're doing an incremental sync. So the
majority of changes are for performing an initial sync, and then saving
the next_batch token to use with SyncUntil.

Otherwise there was a typo fix, inlining the knockingUserID var, and
using must.NotError instead of manual err checking and logging.

* Wait for the kick to complete via sync before knocking again

* Switch out all instances of xyz.amorgan.knock with a var

* const is a thing

* Add instructions on how to stabilise

* Sync after every time we knock

* Allow knocking without a reason in the helper. Add more sync code