Add linting via golangci-lint (#52)
This PR adds a `.golangci.yml` file which the [golangci-lint](https://github.com/golangci/golangci-lint) tool can make use of to run linting over the codebase. This tool was chosen as it is can config and run a multitude of linters and is also the one that is used in [Dendrite](https://github.com/matrix-org/dendrite)'s codebase.