Use server_name query parameters when joining rooms during blueprint exec (#56)
Fixes #38.

This PR allows federating homeservers to create and join other's rooms during blueprint execution. This originally did not work for homeservers which did not extract a `server_name` from a Room ID during room join, such as Synapse.

We now accomplish this by storing the server's name tied to the room's reference when creating the room, and when joining the room look up that server name using the reference.