Handle transactions sent during UnverifiableEvents test
Once again Synapse likes to send presence when it joins a room. I wonder if we
should perhaps make federation.HandleTransactionRequests default for new
servers, with the ability to override it if desired?

The only thing I worry about there is that you may want a test that 404's on
transactions.  Maybe at that point we could just extend
HandleTransactionRequests to return custom responses.