Fixes to Complement's Federation Certificate (#69)
This PR contains two fixes, separated by commit.

1. Add Subject information to Complement's CA and ephemeral certificates. Federation clients will not accept a certificate if they are unable to get the details of the CA that issued them. It looks for this information in the longterm CA cert.
2. The longterm cert would expire after 5 hours. It is not regenerated if it already exists. Federation clients will not accept expired certificates. I've updated the expiration to 10 years.

Because the longterm CA certificate is not regenerated if it already exists, people will need to clear the old certificates (located in `/tests/ca/`) to generate a new certificate.

Note: I think for 1. information is only required for the CA cert, but I've added the same to the ephemeral cert just in case.