Speed up Complement runs
Every `Deploy()` call used to create a network and the deploy the containers.
It took 4s per network and 4s for containers, so 8s for single HSes, 12s for
federated deployments. Both the networks and containers got torn down at the
end of the test via `Destroy()`.

We don't need to re-create networks all the time. It is sufficient to just
make the network once (similar to how we execute blueprints once per run)
and then clean everything up at the *end of all the tests*. This PR does
just that.

Back of the envelope calculations:
```
Number of Deploy() calls: 23
Number of unique blueprints (hence networks): 4
Number of repeat Deploy() calls: 23-4 = 19

Savings per full Complement run: 19*4 = 76s
```