Add potential history visibility test
This test does not work because redacted versions of the events
are backfilled which include the event_id and origin_server_ts
necessary to find it. We just can't see the contents of the message.
So it's ok that this test fails.

See https://github.com/matrix-org/synapse/pull/9445#discussion_r756013164