Send marker events in all federation tests
The marker event is important to let the remote homeservers
know there are historical messages at the given insertion
event it points to.

It's just a lucky happen-stance that the tests passed before
without the need for the marker event.

Marker events are required/expected in order for the history
to be properly visible on remote homservers.