Don't listen on :8448 for the Complement federation server (#289)
* Don't listen on :8448 for the Complement federation server

Instead, listen on a random OS-allocated high numbered port then
do a switcheroo on the `ServerName` so it reads correctly e.g
`host.docker.internal:56185`. This means the server name will be
invalid if it is read before `Server.Listen()` is called so we now
guard common access points in `Server` which rely on the server
name and fail tests if the server is not yet listening when those
functions are called. To further guard against misuse of server
name whilst it isn't valid, turn it into a private field.

* Bind to localhost

* Listen on all interfaces again as we need it to listen on Docker interfaces in CI

* Experimental support for Complement on ubuntu VM in GHA

* Tweak

* More tweaking

* More tweaks

* More

* Guessing at this point

* Maybe

* Modify GOROOT

* Set go version before running other complement install commands

* Docstrings and BUILDKIT=1

* No need to cd complement-master

* Add GOPATH

* Install libolm-dev; more docs

* Say we're not running under CI

* Remove CI flag

We don't need it anymore as CI does not run Complement inside Docker

* Shadow lint

* Fix shadowing