Initial test for partial-state room joins (#362)
A first cut at a test which does a partial-state room join, as proposed in MSC2775. It joins the room, and checks that /sync blocks until the state resyncs (and then checks that the resultant state is sensible).

It won't do anything unless the faster_joins build tag is enabled.