msc2836: fix flakey test caused by /send returning 404
This test uses a `Waiter` to wait for an incoming request from the HS.
If that request doesn't arrive in 1 second it fails the test. Previously,
we had UnexpectedRequestsAreErrors=false in this test to not treat requests
to /send as a failure. However, this would still 404 the /send endpoint. This
would cause Dendrite to begin backing off that server (starting with a 1s backoff)
which meant the Waiter would expire before /event_relationships would be sent.

The fix is simple: 200 OK /send transactions.