Enable arbitrary env vars to be propagated to Complement containers
Fixes https://github.com/matrix-org/complement/issues/6

This PR adds a new complement env var called `COMPLEMENT_SHARE_ENV_PREFIX`.
This can be set to any string e.g `FOO_`. If it is set, Complement will
look at the host environment for any env var with this prefix. If it finds
it, it will remove the prefix and add it to the Complement container.

For example, with a host env:
```
COMPLEMENT_SHARE_ENV_PREFIX=FOO_
FOO_DENDRITE_TRACE_HTTP=1
```
This will cause `DENDRITE_TRACE_HTTP=1` to be set in the Complement container.

This addresses 2 desires:
 - The ability to ad-hoc set debugging env vars when a test is failing
   e.g to use `os.Setenv("DENDRITE_TRACE_HTTP", "1")` in a test to get more logging.
 - The ability to pass through arbitrary data to a Synapse container, as requested
   by @reivilibre