Faster room joins: Fix bug in `awaitPartialStateJoinCompletion` (#487)
`/sync` is not a good way to wait for a room to transition to full
state, since it only waits for the full state at particular events
instead of the whole room. This caused flakiness in the "Device list
updates no longer reach departed servers after partial state join
completes" test.

Use `/members` instead to wait for partial state joins to complete.