Protect CSAPIClients in the HomeserverDeployments structure from concurrent writes.
Appending to a slice is not thread-safe. The actual value is safe, but the header for the slice is
not. Protect it with a Mutex. This will allow more liberal use of t.Parallel() without data races
using RegisterUser(), LoginUser() and Client() to access users on a Homeserver under test when
sharing a deployment.