Faster room joins: Deflake events-arriving-before-prev_events test again (#602)
We previously attempted to deflake this test in #490 by waiting longer
for the 100 outlier events to be persisted. This turned out to be
insufficient, as calculating the full state for all 100 events can also
take a while and delay the room's appearance in /sync.

Raise the timeout for waiting for the room to appear in /sync from 5
seconds to 20 seconds. Going off the logs of the failing test runs, it
takes 6 to 10 seconds for the room to appear in /sync.