Increase thread safety on variables (#592)
* Protect CSAPIClients in the HomeserverDeployments structure from concurrent writes.

Appending to a slice is not thread-safe. The actual value is safe, but the header for the slice is
not. Protect it with a Mutex. This will allow more liberal use of t.Parallel() without data races
using RegisterUser(), LoginUser() and Client() to access users on a Homeserver under test when
sharing a deployment.

* In the client package, the structure CSAPI has a txnID that can be accessed concurrently, mainly by SendEventUnsynced. Make it atomic(fix it in the other places it is used as well).

* In the federation package, the ServerRoom structure has two non-thread-safe parts, State and Timeline. Protect them with a RWMutex,

as most of their usages are reads. This touched the SendJoinRequestHandler as well, as it reads from the State map.

* In msc2716_test.go there is a variable, batchCount, that was causing a data race. Protect it with atomics.

* In room_timestamp_to_event_test.go there is a variable, txnCounter, that was causing a data race. Protect it with atomics.

* In the HomeserverDeployment structure is a map of AccessTokens, protect it with a RWMutex as RegisterUser() can write concurrently when using t.Parallel().

* Make generateLocalpart, a helper function in device_lists_test.go, thread-safe by using atomics.