Add `MustJoinRoom` option to allow Complement to do a partial state join
Some of the faster joins tests have two Complement homeservers in the
same room. We need both Complement homeservers to believe they are in
the room, so that we can wait to observe a leave for the test user as
part of test cleanup. The easiest way to do this is to have the second
Complement homeserver perform a real join.

Signed-off-by: Sean Quah <seanq@matrix.org>