Use setup-go action
This was my first attempt to fix CI. Go 1.17 is no longer available on
runners (https://github.com/actions/runner-images/issues/7276 maybe)
and I wondered if that might help with that.

Unfortunately I think not. But we should get some go build caching as a
side effect of using this action.