Fix CI by ensuring `gotestfmt` is on PATH, plus cleanups (#634)
* Use setup-go action

This was my first attempt to fix CI. Go 1.17 is no longer available on
runners (https://github.com/actions/runner-images/issues/7276 maybe)
and I wondered if that might help with that.

Unfortunately I think not. But we should get some go build caching as a
side effect of using this action.

* Bump checkout action

Drive-by fix for https://github.com/matrix-org/synapse/issues/14203

* Check that gotestfmt is installed

no `-version` flag, so `help` will have to do

* `go install gotestfmt` instead of `go get`

IDK why this suddenly started failing, but `go install` seems to put it
on the PATH in a way that makes CI work.