Add should package
The `should` package is a more lenient form of the `must` package, in
that it does not fail the test if the assertion fails. Instead, these
functions return an error, which the test can then handle.

This is particularly useful in cases where a test needs to retry a
certain request. Previously, you had to hand-roll the checks on the
response because using `must.MatchResponse` would fail the test, even
though it might have been okay as the next retry may pass the matchers.

Fixes https://github.com/matrix-org/complement/issues/546

Tests need to be revisited to see if this can be used in more places.