Custom `Deployment` quality of life improvements (pass `t`, config) (#778)
Spawning from a real use-case with a custom `Deployment`/`Deployer` (Element-internal). I think this is the first working example of a custom deployment so it makes sense that we ran into a couple oversights.

This PR introduces some quality of life for working with a custom `Deployment`. For reference, custom deployments were introduced in https://github.com/matrix-org/complement/pull/750.

 - Pass `t` to allow the custom deployment creation callback (`complement.WithDeployment(callback)`) to handle errors gracefully
 - Also pass in `complement.Config` so the custom Deployment can align behavior to what's configured