Batch up all errors of a room before logging
This allows us to collect all incorrect data about a room and print it.
A side-effect is that one doesn't need to re-run the tests to see all
the broken fields.

In addition, the `must.MatchGJSON` was changed to a `should`, which now
re-polls `/publicRooms`. This prevents the test from exiting early upon
receiving an entry that's missing some keys.