Test `/messages` backfill over federation (#816)
Spawning from wanting to give an accurate reply on https://github.com/element-hq/synapse/issues/19153

But this is just general missing test coverage that's good to have.

There are many more scenarios to test here but Synapse doesn't behave well so it's tough to introduce something that will just fail. Something for the future :fast_forward: